% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AfromMarkers.R
\name{AfromMarkers}
\alias{AfromMarkers}
\title{A estimation from marker gene list}
\usage{
AfromMarkers(data, MGlist, scaleRecover = TRUE)
}
\arguments{
\item{data}{A data set that will be internally coerced into a matrix.
Each row is a gene and each column is a sample.
data should be in non-log linear space with non-negative numerical values (i.e. >= 0).
Missing values are not supported.}

\item{MGlist}{A list of vectors, each of which contains known markers
and/or CAM-detected markers for one subpopulation.}

\item{scaleRecover}{If TRUE, scale ambiguity of each column vector
in A matrix is removed based on sum-to-one constraint on each row.}
}
\value{
returns the estiamted proportion matrix.
}
\description{
This function estimate proportion matrix based on marker gene list and observed
mixture expression data.
}
\details{
On the basis of the expression levels of subpopulation-specific marker genes
    detected by CAM or from literatures, the relative proportions
    of constituent subpopulations are estimated by spatial median using \code{\link[pcaPP]{l1median}}.
    Scale ambiguity is removed optionally.
}
\examples{
#obtain data and marker genes
data(ratMix3)
S <- ratMix3$S
pMGstat <- MGstatistic(S, c("Liver","Brain","Lung"), cores = 30)
pMGlist.FC <- lapply(c("Liver","Brain","Lung"), function(x)
rownames(pMGstat)[pMGstat$idx == x & pMGstat$OVE.FC > 10])

#estimate A matrix from markers
Aest <- AfromMarkers(ratMix3$X, pMGlist.FC)
}
