% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CAMMGObj-class}
\alias{CAMMGObj-class}
\alias{CAMMGObj}
\title{Class "CAMMGObj"}
\description{
An S4 class for storing marker gene detection results.
}
\section{Slots}{

\describe{
\item{\code{idx}}{Two numbers which are two solutions' ranks by sum of
margin-of-error.}

\item{\code{corner}}{The indexes of clusters as detected corners. Each row is a
solution.}

\item{\code{error}}{Two rows. The first row is sum of margin-of-errors for nComb
possible combinations. The second row is reconstruction errors for
nComb possible combinations.}
}}

