% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CAMObj-class}
\alias{CAMObj-class}
\alias{CAMObj}
\title{Class "CAMObj"}
\description{
An S4 class for storing results of CAM.
}
\section{Slots}{

\describe{
\item{\code{PrepResult}}{An object of class "\code{\link{CAMPrepObj}}" storing data
preprocessing results from \code{\link{CAMPrep}} function.}

\item{\code{MGResult}}{A list of "\code{\link{CAMMGObj}}" objects
storing marker gene detection
results from \code{\link{CAMMGCluster}} function for each candidate
subpopulation number.}

\item{\code{ASestResult}}{A list of "\code{\link{CAMASObj}}" objects storing
estimated proportions, subpopution-specific expressions and mdl values
from \code{\link{CAMASest}} function for each candidate
subpopulation number.}
}}

