% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CAMPrepObj-class}
\alias{CAMPrepObj-class}
\alias{CAMPrepObj}
\title{Class "CAMPrepObj"}
\description{
An S4 class for storing data preprocessing results.
}
\section{Slots}{

\describe{
\item{\code{Valid}}{logical vector to indicate the genes left after filtering.}

\item{\code{Xprep}}{Preprocessed data matrix.}

\item{\code{Xproj}}{Preprocessed data matrix after perspective projection.}

\item{\code{W}}{The matrix whose rows are loading vectors.}

\item{\code{cluster}}{cluster results including two vectors.
The first indicates the cluster to which each gene is allocated.
The second is the number of genes in each cluster.}

\item{\code{c.outlier}}{The clusters with the gene number smaller than
MG.num.thres.}

\item{\code{centers}}{The centers of candidate corner clusters (candidate clusters
containing marker genes).}
}}

