% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cornerSort.R
\name{cornerSort}
\alias{cornerSort}
\title{Candidate combinations as corners}
\usage{
cornerSort(X, K, nComb)
}
\arguments{
\item{X}{A matrix of data. Each column is a data point.}

\item{K}{The number of corner points.}

\item{nComb}{The number of possible combinations of data points as corners.}
}
\value{
A list containing the following components:
\item{idx}{A matrix to show the indexes of data points in combinations to construct a convex hull.
Each column is one combination.}
\item{error}{A vector of margin errors for each combination}
}
\description{
Given a set of data points, return possible combinations of data points as corners.
These combinations are selected by ranking margin errors.
}
