% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MDLObj-class}
\alias{MDLObj-class}
\alias{MDLObj}
\title{Class "MDLObj"}
\description{
An S4 class for storing mdl values.
}
\section{Slots}{

\describe{
\item{\code{K}}{The candidate subpopulation numbers.}

\item{\code{datalengths}}{For each model with a certain subpopulation number,
code length of data under the model.}

\item{\code{mdls}}{mdl value for each model with a certain subpopulation number.}
}}

