% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexplot.R
\name{simplexplot}
\alias{simplexplot}
\title{The plot of scatter simplex}
\usage{
simplexplot(data, A, MGlist = NULL, corner.order = NULL,
  data.col = "gray", corner.col = "red", ...)
}
\arguments{
\item{data}{A data set that will be internally coerced into a matrix.
Each row is a gene and each column is a sample.
data should be in non-log linear space with non-negative numerical values (i.e. >= 0).
Missing values are not supported.}

\item{A}{Prior/Estimated proporiton matrix.}

\item{MGlist}{A list of vectors, each of which contains known markers
and/or CAM-detected markers for one subpopulation.}

\item{corner.order}{The order to show simplex corners counterclockwise.}

\item{data.col}{The color for data points. The default is "gray".}

\item{corner.col}{The color for corner points. The default is "red".}
}
\description{
This function shows simplex
}
\details{
This function can show the scatter simplex and deteced marker genes
in a 2D plot. The corners in the high-dimenisonal simplex will still locate
at extreme points of low-dimensional simplex. These corners will follow the
order set by \code{corner.order} to display in the plot counterclockwise.
}
\examples{
simplexplot(data, Aest) #plot simplex for data
simplexplot(data, Aest, MGlist) # Color marker genes in simplex plot

#set differnt corner order and colors
simplexplot(data, Aest, MGlist, corner.order = c(2,1,3),
data.col = "blue", corner.col = c("red","orange","green"))
}
